- Clean main file - Move Background task logic to seperate class. - Re-write bootstrap logic for Workmanager initialization and task schedule
